% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modification.R
\name{add_area_field}
\alias{add_area_field}
\title{Add and calculate an area field for polygon sf objects}
\usage{
add_area_field(sf_obj, field_name)
}
\arguments{
\item{sf_obj}{An polygon sf object}

\item{field_name}{character. The name of the area field.}
}
\value{
Modified sf object
}
\description{
Add and calcualte an area field for polygon sf objects.  Area is calcuated
based on the linear unit of the assigned projection of the input sf object.
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
