% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calulate_polygon_part_area}
\alias{calulate_polygon_part_area}
\title{Calculate the area of the intersected polygon portion}
\usage{
calulate_polygon_part_area(fraction, xmax, coords, ply, total_area)
}
\arguments{
\item{fraction}{The fraction of area being processed represented from 0.0 to 1.0}

\item{xmax}{A float or integer value for the x (longitude) coordinate maximum}

\item{coords}{A list containing xmin, ymin, and ymax}

\item{ply}{An sf polygon object}

\item{total_area}{total area of the whole polygon}
}
\value{
the area of the polygon portion
}
\description{
Calculate the area of the intersected polygon portion
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
