% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{chopped_area}
\alias{chopped_area}
\title{Calculate the area of a slice of a polygon as numeric}
\usage{
chopped_area(xmax, coords, ply)
}
\arguments{
\item{xmax}{A float or integer value for the x (longitude) coordinate maximum}

\item{coords}{A list containing xmin, ymin, and ymax}

\item{ply}{An sf polygon object}
}
\value{
numeric area of a polygon slice
}
\description{
Calculate the area of a slice of a polygon as numeric
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
