% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDAL_functions.R
\name{convert_file}
\alias{convert_file}
\title{convert_file}
\usage{
convert_file(
  inupt_file = "input.tif",
  out_file = "output.bil",
  output_format = "ENVI"
)
}
\arguments{
\item{inupt_file}{location of input file}

\item{out_file}{Name of output file}

\item{output_format}{Format of output. Currently set to ENVI.}
}
\description{
Convert files to specific formats
}
\author{
Kanishka Narayan (kanishka.narayan@pnnl.gov)
}
