% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modification.R
\name{fill_polygon_holes}
\alias{fill_polygon_holes}
\title{Fill holes in polygon by area threshold}
\usage{
fill_polygon_holes(sf_object, sqkm_threshold)
}
\arguments{
\item{sf_object}{sf polygon object}

\item{sqkm_threshold}{area threshold to fill holes smaller than in square kilometers}
}
\description{
Fill holes in polygon by area threshold
}
\details{
Fill holes in polygon by area threshold in square kilometers
}
