% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGIS-Package.R
\name{GeoCoder}
\alias{GeoCoder}
\title{GeoCoder (using Google maps)}
\usage{
GeoCoder(Data, AddressVar, api_key)
}
\arguments{
\item{Data}{a dataframe that contains addresses. It may also contain additional columns.}

\item{AddressVar}{variable that contains addresses}

\item{api_key}{Google's geocode API key}
}
\description{
This function allows user to enter a list of Addresses and use Google maps to locate the coordinate (WGS) of each address. This function requires a valid Google Geocoding API key.
}
\examples{
GeoCoder(AddressDF, Addresses, "XXX")
}

