\name{geodata.adm}
\alias{geodata.adm}
\docType{data}
\title{Geographic data}
\description{ Geographic data
}
\usage{ }
\format{
  SpatialPolygonsDataFrame, in .RData format
}

\references{ \url{http://biogeo.berkeley.edu/gadm/} }

\examples{
\dontrun{
	x <- read.adm()
	plot(x)
}
}
\keyword{datasets}
