% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_intersection_fractions}
\alias{get_intersection_fractions}
\title{Return intersection areas of two spatial files along with proportions used to perform area based downscaling/aggregations}
\usage{
get_intersection_fractions(
  shpfile_1 = "temporary_output/tl_2019_us_state.shp",
  shpfile_2 = "temporary_output/glu_boundaries_moirai_combined_3p1_0p5arcmin.shp",
  area_field_1 = "state_area",
  area_field_2 = "basin_area",
  default_crs = 4326,
  out_csv = NULL,
  out_shape_file = "gcamusa_state_glu_intersections.shp",
  shape_file_cols = c("glu_id", "NAME", "GEOID"),
  write_other_files = FALSE
)
}
\arguments{
\item{shpfile_1}{character. A full path with file name and extension to the input polygon shapefile}

\item{shpfile_2}{character. A full path with file name and extension to the input polygon shapefile}

\item{area_field_1}{character.Name of column with area calculated for shapefile 1}

\item{area_field_2}{character.Name of column with area calculated for shapefile 2}

\item{out_csv}{character. A full path with file name and extension for the output CSV file}

\item{shape_file_cols}{vector. The names to be selected when returning the shape file.}

\item{to_crs}{int. The EPSG number of a projected coordinate reference system.  Must be projected
and not geographic due to assumptions that sf::st_intersection makes when using
geographic systems.  Default is EPSG:4326, WGS 84}

\item{shape_file}{character. The name of the shapefile to be returned}
}
\value{
data.frame
}
\description{
Intersect two shape files/spatial files, return proportions used to downscale or aggregate data based on calculated areas.
The function will return a sf data frame with additional columns including the areas for the spatial boundaries
and the intersection. Two ratios are returned which can be used to downscale data or aggregate data as required.
}
\author{
Kanishka Narayan (kanishka.narayan@pnnl.gov)
}
