% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_raster_val_classes}
\alias{get_raster_val_classes}
\title{Get raster frequency per class from polygon input areas}
\usage{
get_raster_val_classes(raster_object, polygon)
}
\arguments{
\item{raster_object}{character. An object of class RasterLayer.}

\item{polygon}{character. A polygon to define spatial boundary of raster value counts (e.g. a given city's watersheds)}
}
\value{
table of crop types present and their frequency of occurrence
}
\description{
Get the frequency per class of raster values represented in the input raster dataset
when restricted to the input watershed polygons for a target city.  This ignores NA.
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
