% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_raster_val_classes_byslice}
\alias{get_raster_val_classes_byslice}
\title{Get the landclasses intersecting a source polygon for large polygons intersecting a
high-resolution raster.}
\usage{
get_raster_val_classes_byslice(ply, rast, n_parts, n_cores)
}
\arguments{
\item{ply}{An sf polygon object}

\item{rast}{A raster object}

\item{n_parts}{integer. The number of equal area parts to split the polygon into for processing.}
}
\value{
new polygon slice that is next in sequence
}
\description{
This function is to be used on very large polygons that would not generally
run well when harvesting the information contained in the source high-resolution raster.
The source polygon is split into a user-defined number of equal parts and then each of
those parts are used to extract the land classes from the source raster in parallel.
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
