% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contours.R
\name{getSpatialContours}
\alias{getSpatialContours}
\title{Get spatial contours}
\usage{
getSpatialContours(xyz, contour_values = seq(from = -10, to = -200, by = -20))
}
\arguments{
\item{xyz}{A dataframe with column x, y, z data (long, lat, alt)}

\item{contour_values}{A numeric vector of z values to contour. If this argument is provided, the stacking order of the contours will be preserved in the order of first occurence within the supplied vector.}
}
\value{
Returns a SpatialLinesDataFrame object (sp package) that can be used to draw all contour lines. All contour lines with a certain z-value can be found with 'contour_value' in the df.
}
\description{
Get spatial contours
}
