% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_ncdf_to_raster}
\alias{import_ncdf_to_raster}
\title{Import NetCDF to brick raster}
\usage{
import_ncdf_to_raster(ncdf_file)
}
\arguments{
\item{ncdf_path}{character. A full path to the input NetCDF file with the file name and extension}
}
\value{
raster object
}
\description{
Import NetCDF to brick raster
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
