% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_raster}
\alias{import_raster}
\title{Create a raster object from a file path}
\usage{
import_raster(raster_path)
}
\arguments{
\item{raster_path}{character. A full path to the input raster file with file name and extension}
}
\value{
raster object
}
\description{
Create a raster object from a full path to raster file name and extension
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
