% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_shapefile}
\alias{import_shapefile}
\title{Create an sf object from a shapefile}
\usage{
import_shapefile(shp_path, quiet = TRUE, method = "sf")
}
\arguments{
\item{shp_path}{character. A full path to the input shapefile with file name and extension}

\item{quiet}{boolean.}

\item{method}{tool used to convert to spatial data; either 'sf' or 'rgdal'}
}
\value{
sf object
}
\description{
Create an sf object from a full path to shapefile with file name and extension
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
