% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{mask_raster_to_polygon}
\alias{mask_raster_to_polygon}
\title{Mask raster to polygon}
\usage{
mask_raster_to_polygon(raster_object, polygon)
}
\arguments{
\item{raster_object}{character. An object of class RasterLayer.}

\item{polygon}{character. A polygon to define spatial boundary of raster value counts (e.g. a given city's watersheds)}
}
\description{
Mask raster to polygon
}
\details{
masks a raster file against a chosen polygon.
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
