% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ncdf_to_csv}
\alias{ncdf_to_csv}
\title{Convert NetCDF file to a CSV}
\usage{
ncdf_to_csv(ncdf_file, out_csv, nc_var_name, resolution = 0.5,
  start_year = NULL, through_year = NULL, nmonths = NULL)
}
\arguments{
\item{ncdf_file}{character. asdf}

\item{out_csv}{character. asdf}

\item{nc_var_name}{character. asdf}

\item{resolution}{float. asdf}

\item{start_year}{int}

\item{through_year}{int}

\item{nmonths}{int}
}
\description{
Convert NetCDF file to a CSV
}
\author{
Chris Vernon (chris.vernon@pnnl.gov)
}
