% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{poly_intersect}
\alias{poly_intersect}
\title{Create a slice of the original polygon object from a smaller bounding box}
\usage{
poly_intersect(xmin, xmax, ymin, ymax, ply)
}
\arguments{
\item{xmin}{A float or integer value for the x (longitude) coordinate minimum}

\item{xmax}{A float or integer value for the x (longitude) coordinate maximum}

\item{ymin}{A float or integer value for the y (latitude) coordinate minimum}

\item{ymax}{A float or integer value for the y (latitude) coordinate maximum}

\item{ply}{An sf polygon object}
}
\value{
An sf polygon object
}
\description{
Create a slice of the original polygon object from a smaller bounding box
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
