% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{polygon_bounding_box}
\alias{polygon_bounding_box}
\title{Create a spatial polygon bounding box sf object}
\usage{
polygon_bounding_box(x_min, x_max, y_min, y_max, my_crs)
}
\arguments{
\item{x_min}{A float or integer value for the x (longitude) coordinate minimum}

\item{x_max}{A float or integer value for the x (longitude) coordinate maximum}

\item{y_min}{A float or integer value for the y (latitude) coordinate minimum}

\item{y_max}{A float or integer value for the y (latitude) coordinate maximum}

\item{my_crs}{An integer for the EPSG number of the desired output coordinate
reference system.}
}
\value{
A bounding box polygon as an sf object
}
\description{
Creates a spatial polygon bounding box from a user-provided extent
and coordinate reference system.
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
