% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{raster_to_polygon}
\alias{raster_to_polygon}
\title{Create a polygon sf object from a raster}
\usage{
raster_to_polygon(raster, na.rm = FALSE)
}
\arguments{
\item{raster}{A raster, stack, or brick object}

\item{na.rm}{boolean. If TRUE will polygonize only non-NA cells. Defualt is FALSE.}
}
\description{
Create a polygon sf object from a raster, stack, or brick.
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
