% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGIS-package.R
\docType{package}
\name{RGIS-package}
\alias{RGIS}
\alias{RGIS-package}
\title{RGIS: visualizador de camadas geoespaciais}
\description{
O RGIS abre um mapa interativo no navegador e permite carregar ocorrencias em
CSV, rasters continuos e camadas vetoriais, reprojetando tudo para WGS84.

Ponto de entrada: \code{\link[=run_app]{run_app()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diogosbr/RGIS}
  \item Report bugs at \url{https://github.com/diogosbr/RGIS/issues}
}

}
\author{
\strong{Maintainer}: Diogo Rocha \email{d.rocha@iis-rio.org}

Authors:
\itemize{
  \item Diogo Rocha \email{d.rocha@iis-rio.org}
}

}
\keyword{internal}
