% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgis.R
\docType{package}
\name{rgis}
\alias{rgis}
\alias{rgis-package}
\title{rgis:  R-based Geographic Information System (GIS) utilities}
\description{
The \code{rgis} package facilitates GIS functionality and workflows
commonly reprented in propritary and open source GISs. This package
does not contain functionality for visulization, but is focsed on
geospatial algorithms for analysis, conversion, IO, modification, and
workflows of spatial files and data structures.
}
