% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{slicer}
\alias{slicer}
\title{Create next polygon slice in sequence}
\usage{
slicer(ply, xmin, xmax)
}
\arguments{
\item{ply}{An sf polygon object}

\item{xmin}{A float or integer value for the x (longitude) coordinate minimum}

\item{xmax}{A float or integer value for the x (longitude) coordinate maximum}
}
\value{
new polygon slice that is next in sequence
}
\description{
Create next polygon slice in sequence
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
