% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{warp_raster}
\alias{warp_raster}
\title{warp_raster}
\usage{
warp_raster(
  input_file = "input.tif",
  input_format = ".tif",
  NO_DATA_STATUS = TRUE,
  method = "average",
  no_data_value = 127,
  resolution = "+proj=longlat +ellps=WGS84",
  col_row = c(4320, 2160),
  extent = c(-180, -90, 180, 90),
  out_file = "outfile",
  output_format = "ENVI"
)
}
\arguments{
\item{input_format}{format of the input file}

\item{NO_DATA_STATUS}{Boolean for no_data values. Set this to true if there are NO_DATA values in the data set. If there are no data values, this will fill them with 0}

\item{method}{Method of conversion. Default is set to average}

\item{no_data_value}{Value of no_data in the dataset}

\item{col_row}{columns and rows in output dataset. Defualt is set to c(4320,2160)}

\item{extent}{output extent. Default is set to c(-180,-90,180,90)}

\item{out_file}{Name of output file}

\item{output_format}{Format of output. Currently set to ENVI.}

\item{inupt_file}{location of input file}

\item{projection}{Output projection type. Defualt is set to +proj=longlat +ellps=WGS84}
}
\description{
Convert raster to specific dimensions and formats
}
\author{
Kanishka Narayan (kanishka.narayan@pnnl.gov)
}
