% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgis-package.R
\docType{package}
\name{rgis-package}
\alias{rgis}
\alias{rgis-package}
\title{rgis package}
\description{
Functions to perform GIS operations in R.
}
\author{
Francisco Rodriguez-Sanchez
}
\keyword{package}

