\name{Rgis-package}
\alias{Rgis-package}
\alias{Rgis}
\docType{package}
\title{ Geographic Data for mapping and analysis
}
\description{
Download (if necessary) and read geographic data into a standard format for mapping and anaylis
}
\details{
	\tabular{ll}{
		Package: \tab Rgis\cr
		Type: \tab Package\cr
		Version: \tab 0.1\cr
		Date: \tab 2008-10-07\cr
		License: \tab GPL v3\cr
		LazyLoad: \tab yes\cr
}

Overview here.....
}
\author{
Robert J. Hijmans \email{r.hijmans@gmail.com}
Jacob van Etten

Maintainer: Robert J. Hijmans \email{r.hijmans@gmail.com}
}
\references{

}
\keyword{ package }
\keyword{ spatial }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

\examples{
}
